# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: MFK — Smart Vehicle Diagnostics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /diagnostics/sessions:
    get:
      operationId: listDiagnosticSessions
      tags: [diagnostics]
      parameters:
        - in: query
          name: vehicleId
          schema: { type: string }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DiagnosticSession"
    post:
      operationId: startDiagnosticSession
      tags: [diagnostics]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartSessionBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSession"
  /diagnostics/sessions/{sessionId}:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema: { type: string }
    get:
      operationId: getDiagnosticSession
      tags: [diagnostics]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSessionDetail"
  /diagnostics/sessions/{sessionId}/close:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema: { type: string }
    patch:
      operationId: closeDiagnosticSession
      tags: [diagnostics]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiagnosticSession"
  /diagnostics/sessions/{sessionId}/telemetry:
    parameters:
      - in: path
        name: sessionId
        required: true
        schema: { type: string }
    get:
      operationId: getSessionTelemetry
      tags: [diagnostics]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TelemetryPoint"
...
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
  /bookings/{bookingId}:
    parameters:
      - in: path
        name: bookingId
        required: true
        schema: { type: string }
    patch:
      operationId: updateBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"

  # ========== AI ASSISTANT ==========
  /ai/chat:
    post:
      operationId: aiChat
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
  /ai/recommendations/{vehicleId}:
    parameters:
      - in: path
        name: vehicleId
        required: true
        schema: { type: string }
    get:
      operationId: getAiRecommendations
      tags: [ai]
      summary: Predicted issues + suggested maintenance for a vehicle
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
...
components:
  schemas:
...
    DiagnosticSession:
      type: object
      required: [id, vehicleId, startedAt, status]
      properties:
        id: { type: string }
        vehicleId: { type: string }
        vehicleNickname: { type: string, nullable: true }
        vehicleMake: { type: string }
        vehicleModel: { type: string }
        startedAt: { type: string, format: date-time }
        endedAt: { type: string, format: date-time, nullable: true }
        durationSec: { type: integer, nullable: true }
        odometerKm: { type: integer, nullable: true }
        dtcCount: { type: integer }
        healthBefore: { type: integer, nullable: true }
        healthAfter: { type: integer, nullable: true }
        status:
          type: string
          enum: [active, completed]
    DiagnosticSessionDetail:
      allOf:
        - $ref: "#/components/schemas/DiagnosticSession"
        - type: object
          properties:
            dtcCodes:
              type: array
              items:
                $ref: "#/components/schemas/DtcCode"
            telemetrySummary:
              $ref: "#/components/schemas/TelemetrySummary"
    StartSessionBody:
      type: object
      required: [vehicleId]
      properties:
        vehicleId: { type: string }
        odometerKm: { type: integer }
    TelemetryPoint:
      type: object
      required: [time]
      properties:
        time: { type: string, format: date-time }
        rpm: { type: integer, nullable: true }
        speedKmh: { type: integer, nullable: true }
        coolantTemp: { type: integer, nullable: true }
        intakeTemp: { type: integer, nullable: true }
        batteryV: { type: number, nullable: true }
        fuelLevelPct: { type: integer, nullable: true }
        engineLoad: { type: integer, nullable: true }
        throttlePos: { type: integer, nullable: true }
...
    TelemetrySummary:
      type: object
      properties:
        avgRpm: { type: number, nullable: true }
        maxSpeed: { type: integer, nullable: true }
        avgCoolantTemp: { type: number, nullable: true }
        minBatteryV: { type: number, nullable: true }
        sampleCount: { type: integer }

    # ===== DTC =====
    DtcCode:
      type: object
      required: [id, code, status, severity, vehicleId]
      properties:
        id: { type: string }
        sessionId: { type: string, nullable: true }
        vehicleId: { type: string }
        vehicleMake: { type: string, nullable: true }
        vehicleModel: { type: string, nullable: true }
        code: { type: string }
        status:
          type: string
          enum: [active, cleared, pending]
        severity:
          type: string
          enum: [low, medium, high, critical]
        descriptionEn: { type: string, nullable: true }
        descriptionAr: { type: string, nullable: true }
        possibleCauses:
          type: array
          items: { type: string }
        estimatedCostMin: { type: integer, nullable: true }
        estimatedCostMax: { type: integer, nullable: true }
        recommendedAction:
          type: string
          nullable: true
          enum: [drive_now, schedule_week, monitor]
        actionReasonAr: { type: string, nullable: true }
        detectedAt: { type: string, format: date-time }
        clearedAt: { type: string, format: date-time, nullable: true }
...
    Booking:
      type: object
      required: [id, vehicleId, workshopId, scheduledAt, status]
      properties:
        id: { type: string }
        vehicleId: { type: string }
        vehicleMake: { type: string }
        vehicleModel: { type: string }
        workshopId: { type: string }
        workshopName: { type: string }
        workshopNameAr: { type: string }
        serviceType: { type: string }
        serviceTypeAr: { type: string }
        scheduledAt: { type: string, format: date-time }
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        estimatedCost: { type: integer, nullable: true }
        finalCost: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
    CreateBookingBody:
      type: object
      required: [vehicleId, workshopId, serviceType, scheduledAt]
      properties:
        vehicleId: { type: string }
        workshopId: { type: string }
        serviceType: { type: string }
        scheduledAt: { type: string, format: date-time }
        notes: { type: string }
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, in_progress, completed, cancelled]
        scheduledAt: { type: string, format: date-time }

    # ===== AI =====
    AiChatBody:
      type: object
      required: [message]
      properties:
        message: { type: string }
        vehicleId: { type: string }
    AiChatResponse:
      type: object
      required: [reply]
      properties:
        reply: { type: string }
        suggestedActions:
          type: array
          items:
            type: object
            properties:
              labelAr: { type: string }
              kind:
                type: string
                enum: [book_workshop, view_dtc, schedule_maintenance, view_vehicle]
              targetId: { type: string }
    Recommendation:
      type: object
      required: [id, vehicleId, kind, severity, titleAr]
      properties:
        id: { type: string }
        vehicleId: { type: string }
        kind:
          type: string
          enum: [predictive_failure, maintenance_due, telemetry_anomaly, behavioral]
        severity:
          type: string
          enum: [info, warning, critical]
        titleAr: { type: string }
        descriptionAr: { type: string }
        confidencePct: { type: integer }
        suggestedAction: { type: string, nullable: true }
        suggestedCostSar: { type: integer, nullable: true }
        createdAt: { type: string, format: date-time }
...